# plugins/cgi: improve performance for santitizing file descriptors