Only close file descriptors that are known to be open when /proc/self/fd
or /dev/fd exists. Fall back and loop over all when those are missing.

This improves performance significantly when number of max fds is high.

fixes #2053