@ncopa 2020-04-20
Difference is:
```
Requests per second:    9.58 [#/sec] (mean)
```
vs
```
Requests per second:    48.31 [#/sec] (mean)
```

Details:
Without fd patch:

```
ab -n 100 -c 10 https://cgit.alpinelinux.org/aports
This is ApacheBench, Version 2.3 <$Revision: 1874286 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking cgit.alpinelinux.org (be patient).....done


Server Software:        nginx/1.17.8
Server Hostname:        cgit.alpinelinux.org
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
Server Temp Key:        X25519 253 bits
TLS Server Name:        cgit.alpinelinux.org

Document Path:          /aports
Document Length:        15138 bytes

Concurrency Level:      10
Time taken for tests:   10.434 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      1534200 bytes
HTML transferred:       1513800 bytes
Requests per second:    9.58 [#/sec] (mean)
Time per request:       1043.399 [ms] (mean)
Time per request:       104.340 [ms] (mean, across all concurrent requests)
Transfer rate:          143.59 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       74   90  15.3     88     221
Processing:   399  883 130.6    908    1075
Waiting:      369  825 124.4    857    1030
Total:        486  972 133.3    996    1204

Percentage of the requests served within a certain time (ms)
  50%    996
  66%   1051
  75%   1070
  80%   1087
  90%   1122
  95%   1145
  98%   1169
  99%   1204
 100%   1204 (longest request)
```

With fd patch:
```
ab -n 100 -c 10 https://cgit.alpinelinux.org/aports
This is ApacheBench, Version 2.3 <$Revision: 1874286 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking cgit.alpinelinux.org (be patient).....done


Server Software:        nginx/1.17.8
Server Hostname:        cgit.alpinelinux.org
Server Port:            443
SSL/TLS Protocol:       TLSv1.2,ECDHE-RSA-AES256-GCM-SHA384,2048,256
Server Temp Key:        X25519 253 bits
TLS Server Name:        cgit.alpinelinux.org

Document Path:          /aports
Document Length:        15138 bytes

Concurrency Level:      10
Time taken for tests:   2.070 seconds
Complete requests:      100
Failed requests:        0
Total transferred:      1534200 bytes
HTML transferred:       1513800 bytes
Requests per second:    48.31 [#/sec] (mean)
Time per request:       206.992 [ms] (mean)
Time per request:       20.699 [ms] (mean, across all concurrent requests)
Transfer rate:          723.82 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       71  114  55.5     97     453
Processing:    55   78  17.4     73     149
Waiting:       26   39   9.0     37      69
Total:        135  192  57.1    176     526

Percentage of the requests served within a certain time (ms)
  50%    176
  66%    187
  75%    195
  80%    208
  90%    251
  95%    336
  98%    348
  99%    526
 100%    526 (longest request)
```