@rdeioris 2020-04-20
Thanks @ncopa, /dev/fd should grant a good amount of *BSD support. @xrmx i do not know if we need to add a bunch of #ifdef as filesystem access/stat is pretty cheap.