@ncopa 2020-04-20
> Thanks @ncopa, /dev/fd should grant a good amount of *BSD support. 

It should also cover MacOS.

> i do not know if we need to add a bunch of #ifdef as filesystem access/stat is pretty cheap.

I don't think `#ifdef`'ing is worth it here. There are BSDs with linux compat and it appears that linux also has `/dev/fd`.

We could do something like:
```c
	DIR *dirp;
#if defined(__linux__)
	dirp = opendir("/proc/self/fd");
￼	if (dirp == NULL)
#endif
		dirp = opendir("/dev/fd");
```
or
```c
	DIR *dirp = NULL;
#if defined(__linux__)
	dirp = opendir("/proc/self/fd");
#endif
￼	if (dirp == NULL)
		dirp = opendir("/dev/fd"); /* BSDs and MacOS */
```
but I would prefer not add lots of different `#ifdef`s.