@xrmx 2020-04-20
I have /dev/fd on debian too, so agree adding #ifdefery is not worthwhile.