@ncopa 2020-08-17
I've recently learned that it is not allowed to call async unsafe functions after fork, which may result in deadlocks. opendir() does malloc so it is not really allowed to call it after fork. We can either do the opendir before fork or we can use poll to speed up the closing of fds.